Recreated the SPARQLStore, and added the specialized UniProtStore. Still missing to define the search method.
